Se crea un caso para cada opcion del menu y se hacen usos de las clases

+++ Momento2/CLASE9/Edicion.java
@@ -9,7 +9,7 @@ public class Edicion extends Producto {
         this.softwareEditor = softwareEditor;
     }
 //Getters
-    public int getHoras() {
+    public double getHoras() {
         return horas;
     }
     public String softwareEditor() {
